+++ tartiflette/types/exceptions/tartiflette.py
@@ -151,6 +151,10 @@ class NotLoneAnonymousOperation(GraphQLError):
     pass
 
 
+class MultipleRootNodeOnSubscriptionOperation(GraphQLError):
+    pass
+
+
 class UnknownVariableException(GraphQLError):
     def __init__(self, varname):
         # TODO: Unify error messages format